+++ src/app/components/accompany/CarouselComponent.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import React from 'react'
 import { Carousel } from 'antd'
